Revert "[cleanup] Delete chaining package as not needed"
This reverts commit 78ef09d0d3a2396030a0513d45c81b316a1af9c6.

The site does not like this removed.  Its clearly a bug with maven.  To keep at least this working well between javax and jakarta namespaces just drop back and skip the site.  The site is not imporant on this repo at the moment and already has an issue to address.